restructured app, added a folder called 'public' for the frontend photo app to live in. Created separate files for each model.
