+++ package.json
@@ -18,10 +18,10 @@
   "homepage": "https://github.com/cdavis27/photo-album-api",
   "dependencies": {
     "body-parser": "^1.12.3",
-    "cors": "^2.6.0",
     "exif": "^0.4.0",
     "express": "^4.12.3",
     "mongoose": "^4.0.2",
+    "morgan": "^1.6.0",
     "multer": "^0.1.8",
     "q": "^1.3.0"
   }
