+++ public/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Nothing to see here</title>
+</head>
+<body>
+
+<h2>Hello! This is where an HTML5 app lives</h2>
+
+
+</body>
+</html>
